use fe/loc.h and fe/sym.h
