+++ thorin/def.h
@@ -10,7 +10,7 @@
 
 #include "thorin/util/array.h"
 #include "thorin/util/hash.h"
-#include "thorin/util/loc.h"
+#include "thorin/util/dbg.h"
 #include "thorin/util/print.h"
 #include "thorin/util/util.h"
