+++ thorin/fe/scopes.h
@@ -2,7 +2,7 @@
 
 #include <deque>
 
-#include "thorin/util/loc.h"
+#include "thorin/util/dbg.h"
 
 namespace thorin {
 
@@ -11,7 +11,7 @@ class Ptrn;
 
 class Scopes {
 public:
-    using Scope = SymMap<std::pair<Loc, const Def*>>;
+    using Scope = fe::SymMap<std::pair<Loc, const Def*>>;
 
     Scopes() { push(); /* root scope */ }
